refactor: move controller to process-example-core
