+++ examples/process-example/process-example-camunda-7/src/main/java/io/miragon/miranum/examples/process/c7/ProcessExampleC7Application.java
@@ -1,12 +1,12 @@
 package io.miragon.miranum.examples.process.c7;
 
-import io.miragon.miranum.examples.process.ExampleProcessConfiguration;
+import io.miragon.miranum.examples.process.ProcessExampleConfiguration;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.context.annotation.Import;
 
 @SpringBootApplication
-@Import(ExampleProcessConfiguration.class)
+@Import(ProcessExampleConfiguration.class)
 public class ProcessExampleC7Application {
 
     public static void main(final String[] args) {
