+++ examples/process-example/process-example-core/pom.xml
@@ -19,11 +19,19 @@
             <groupId>io.miragon.miranum</groupId>
             <artifactId>worker-core</artifactId>
             <version>${project.version}</version>
-            <scope>provided</scope>
         </dependency>
         <dependency>
             <groupId>org.springframework</groupId>
             <artifactId>spring-context</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>org.springframework.boot</groupId>
+            <artifactId>spring-boot-starter-web</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>io.miragon.miranum</groupId>
+            <artifactId>process-core</artifactId>
+            <version>${project.version}</version>
         </dependency>
     </dependencies>
